requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
idna==3.4
pyaes==1.6.1
pycparser==2.21
Pyrogram==2.0.59
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
python-dotenv==0.21.1
python-dotenv-vault==0.6.0
requests==2.31.0
urllib3==2.0.3
